working conditionOnPawn implementation

+++ src/main/java/ro/anud/globalcooldown/model/pawn/PawnOutputModel.java
@@ -7,12 +7,12 @@ import ro.anud.globalcooldown.geometry.Point;
 @Data
 @Builder
 public class PawnOutputModel {
-	private Long id;
-	private Integer value;
-	private String name;
-	private Long version;
-	private Point point;
-	private Long speed;
-	private Long characterCode;
-	private Long userId;
+    private Long id;
+    private Long value;
+    private String name;
+    private Long version;
+    private Point point;
+    private Long speed;
+    private Long characterCode;
+    private Long userId;
 }
