+++ src/main/resources/db/migration/V2__createPawnTable.sql
@@ -4,7 +4,7 @@ CREATE TABLE pawn (
   name           VARCHAR(25) NOT NULL,
   user_id        BIGINT REFERENCES users (id),
   version        BIGINT      NOT NULL DEFAULT 0,
-  val            INT         NOT NULL,
+  val            BIGINT      NOT NULL,
   x              BIGINT      NOT NULL,
   y              BIGINT      NOT NULL,
   speed          BIGINT      NOT NULL,
